1. remove original Classifier class, and rename NoiseAwareModel to Classifier; 2. use the same setting for both binary and multi-class classifier; 3. remove LabelBalancer; 4. code clean up
