+++ src/fonduer/learning/disc_models/sparse_logistic_regression.py
@@ -130,10 +130,8 @@ class SparseLogisticRegression(NoiseAwareModel):
         if "input_dim" not in self.settings:
             raise ValueError("Model parameter input_dim cannot be None.")
 
-        cardinality = self.cardinality if self.cardinality > 2 else 1
-
         self.sparse_linear = SparseLinear(
-            self.settings["input_dim"], cardinality, self.settings["bias"]
+            self.settings["input_dim"], self.cardinality, self.settings["bias"]
         )
 
     def _calc_logits(self, X, batch_size=None):
@@ -173,9 +171,6 @@ class SparseLogisticRegression(NoiseAwareModel):
                 values = values.cuda()
 
             output = self.forward(features, values)
-            if self.cardinality == 2:
-                outputs = torch.cat((outputs, output.view(-1)), 0)
-            else:
-                outputs = torch.cat((outputs, output), 0)
+            outputs = torch.cat((outputs, output), 0)
 
         return outputs
